requirements.txt
asgiref==3.6.0
Django==4.1.7
django-extensions==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.3
